requirements.txt
fastapi
uvicorn
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai
langchain
langchain-google-genai
python-dotenv
